feat: rename deprecated table to tablev4
