--- src/components/Table/index.ts
@@ -1,3 +1,0 @@
-export { default } from './Table';
-export * from './Table';
-export * from './types';
