+++ src/components/TableV4/TableRowContext.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import type { Row, TableType } from './Table';
+import type { Row, TableType } from './TableV4';
 
 export type TableRowContextProps<T extends { [key: string]: unknown }> = {
   row: Row<T>;
