+++ src/components/TableV4/TableV4.mdx
@@ -1,12 +1,12 @@
 import { Meta, Canvas, ArgTypes } from '@storybook/blocks';
-import Table from './Table';
+import TableV4 from './TableV4';
 import { CellStory, RowStory, ExtendedColumnStory } from './utils/TableStoryComponents';
-import * as TableStories from './Table.stories';
+import * as TableStories from './TableV4.stories';
 
 <Meta of={TableStories} />
 
 <SectionHeader
-  title={'Table'}
+  title={'TableV4'}
   isPending
   sections={[
     { title: 'Overview', href: '#overview' },
@@ -22,7 +22,7 @@ A universal Table component that applies Orfium basic usages. Check inline data 
 
 ## Table Props
 
-<ArgTypes of={Table} />
+<ArgTypes of={TableV4} />
 
 ## Cell Type
 
@@ -39,9 +39,9 @@ A universal Table component that applies Orfium basic usages. Check inline data 
 ## Usage
 
 ```js
-import { Table } from '@orfium/ictinus';
+import { TableV4 } from '@orfium/ictinus';
 
-<Table />;
+<TableV4 />;
 ```
 
 <SubsectionHeader title="Variants" />
