+++ src/components/TableV4/index.ts
@@ -1,0 +1,3 @@
+export { default } from './TableV4';
+export * from './TableV4';
+export * from './types';
