+++ src/components/TruncatedContent/TruncatedContent.mdx
@@ -1,6 +1,5 @@
 import { Meta, Canvas, ArgTypes, Story } from '@storybook/blocks';
 import TruncatedContent from './TruncatedContent';
-import Table from '../Table';
 import * as TruncatedContentStories from './TruncatedContent.stories';
 
 <Meta of={TruncatedContentStories} />
